index: add index.html
It is important to notice that this is an old code that I wrote when I
was graduating and It is very messy hehe.

The purpouse of this is to guide people who is learning programming to
make a simple JavaScript game. I implmented this code on a YouTube
tutorial serie and for being consistent with the videos that are
published on my YouTube channel, I decided do not refactor this code.